Move PlaylistTags into its own file
Also add some regions to PlaylistViewModel for folding.
